Thingsee OS SDK update to version <2017-04-04>
Thingsee Engine FW fixes:
 - sense_group_location: return from gps callback if engine
   state changes

Initial State connector:
 - do not send separate latitude/longitude events if merge
   lat/lon mode is enabled

+++ apps/tests/unity_usrsock/defines.h
@@ -107,9 +107,9 @@ int usrsocktest_daemon_get_num_waiting_connect_sockets(void);
 
 int usrsocktest_daemon_get_num_recv_empty_sockets(void);
 
-int usrsocktest_daemon_get_send_bytes(void);
+ssize_t usrsocktest_daemon_get_send_bytes(void);
 
-int usrsocktest_daemon_get_recv_bytes(void);
+ssize_t usrsocktest_daemon_get_recv_bytes(void);
 
 int usrsocktest_daemon_get_num_unreachable_sockets(void);
